:art: FORMAT (reducers): move project reducer to root
